pyproject.toml
[project]
name = "hearpauljennings"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.134.0",
    "faster-whisper>=1.2.1",
    "httpx>=0.28.1",
    "numpy>=2.4.2",
    "openpyxl>=3.1.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "sentence-transformers>=5.2.3",
    "uvicorn>=0.41.0",
]
